requirements.txt
Pillow
pandas
openpyxl
smtplib
python-dotenv
logging
